Fix string reference table with empty string
It seems that ActionScript is putting empty string in string reference
table during encoding/decoding. While the specification states that
"the empty string is never sent by reference" it does not specify if
it should be put in reference table or not.
php-amf3 does not put empty string in reference table also Zend AMF,
written by Adobe, also is not doing this. This however causes it
impossible to decode any data from ActionScript that uses empty
string, since string references stop working.
Fix it by putting empty string in string reference table (but do not
send it as reference).